Backend/requirements.txt
fastapi
uvicorn[standard]
chromadb
sentence-transformers
python-dotenv
sqlalchemy
passlib[bcrypt]